Rename "explorer" folder to "tree"
