+++ src/commands/deleteWebApp.ts
@@ -4,9 +4,9 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { IActionContext } from "vscode-azureextensionui";
-import { SiteTreeItem } from "../explorer/SiteTreeItem";
-import { WebAppTreeItem } from "../explorer/WebAppTreeItem";
 import { ext } from "../extensionVariables";
+import { SiteTreeItem } from "../tree/SiteTreeItem";
+import { WebAppTreeItem } from "../tree/WebAppTreeItem";
 
 export async function deleteWebApp(context: IActionContext, node?: SiteTreeItem): Promise<void> {
     if (!node) {
