+++ src/commands/deploy/deploy.ts
@@ -11,10 +11,10 @@ import * as appservice from 'vscode-azureappservice';
 import { getDeployFsPath, getDeployNode, IDeployContext, IDeployPaths, showDeployConfirmation } from 'vscode-azureappservice';
 import { IActionContext, parseError } from 'vscode-azureextensionui';
 import * as constants from '../../constants';
-import { SiteTreeItem } from '../../explorer/SiteTreeItem';
-import { WebAppTreeItem } from '../../explorer/WebAppTreeItem';
 import { ext } from '../../extensionVariables';
 import { localize } from '../../localize';
+import { SiteTreeItem } from '../../tree/SiteTreeItem';
+import { WebAppTreeItem } from '../../tree/WebAppTreeItem';
 import { javaUtils } from '../../utils/javaUtils';
 import { nonNullValue } from '../../utils/nonNull';
 import { isPathEqual } from '../../utils/pathUtils';
