+++ src/commands/generateDeploymentScript.ts
@@ -8,9 +8,9 @@ import * as fse from 'fs-extra';
 import * as path from 'path';
 import { ProgressLocation, window, workspace } from "vscode";
 import { IActionContext } from "vscode-azureextensionui";
-import { WebAppTreeItem } from "../explorer/WebAppTreeItem";
 import { ext } from "../extensionVariables";
 import { localize } from "../localize";
+import { WebAppTreeItem } from "../tree/WebAppTreeItem";
 import { createResourceClient } from "../utils/azureClients";
 import { nonNullValue } from "../utils/nonNull";
 import { getResourcesPath } from "../utils/pathUtils";
