+++ src/commands/logstream/enableFileLogging.ts
@@ -6,10 +6,10 @@
 import * as vscode from 'vscode';
 import { LogFilesTreeItem } from 'vscode-azureappservice';
 import { DialogResponses, IActionContext } from 'vscode-azureextensionui';
-import { SiteTreeItem } from "../../explorer/SiteTreeItem";
-import { WebAppTreeItem } from '../../explorer/WebAppTreeItem';
 import { ext } from '../../extensionVariables';
 import { localize } from '../../localize';
+import { SiteTreeItem } from "../../tree/SiteTreeItem";
+import { WebAppTreeItem } from '../../tree/WebAppTreeItem';
 
 export interface IEnableFileLoggingContext extends IActionContext {
     suppressAlreadyEnabledMessage?: boolean;
