+++ src/commands/postDeploy/getLinuxDetectorError.ts
@@ -7,8 +7,8 @@ import { ServiceClient } from '@azure/ms-rest-js';
 import * as moment from 'moment';
 import { createGenericClient, IActionContext } from "vscode-azureextensionui";
 import { detectorTimestampFormat } from "../../constants";
-import { SiteTreeItem } from "../../explorer/SiteTreeItem";
 import { localize } from "../../localize";
+import { SiteTreeItem } from "../../tree/SiteTreeItem";
 import { findTableByName, getValuesByColumnName } from "./parseDetectorResponse";
 
 export enum ColumnName {
