+++ src/commands/startSsh.ts
@@ -8,10 +8,10 @@ import * as portfinder from 'portfinder';
 import * as vscode from 'vscode';
 import { reportMessage, setRemoteDebug, SiteClient, TunnelProxy } from 'vscode-azureappservice';
 import { IActionContext } from 'vscode-azureextensionui';
-import { SiteTreeItem } from '../explorer/SiteTreeItem';
-import { WebAppTreeItem } from '../explorer/WebAppTreeItem';
 import { ext } from '../extensionVariables';
 import { localize } from '../localize';
+import { SiteTreeItem } from '../tree/SiteTreeItem';
+import { WebAppTreeItem } from '../tree/WebAppTreeItem';
 import { delay } from '../utils/delay';
 
 export type sshTerminal = {
