+++ src/commands/stopWebApp.ts
@@ -6,10 +6,10 @@
 
 import { SiteClient } from "vscode-azureappservice";
 import { IActionContext } from "vscode-azureextensionui";
-import { SiteTreeItem } from "../explorer/SiteTreeItem";
-import { WebAppTreeItem } from "../explorer/WebAppTreeItem";
 import { ext } from "../extensionVariables";
 import { localize } from "../localize";
+import { SiteTreeItem } from "../tree/SiteTreeItem";
+import { WebAppTreeItem } from "../tree/WebAppTreeItem";
 
 export async function stopWebApp(context: IActionContext, node?: SiteTreeItem): Promise<void> {
     if (!node) {
