+++ src/commands/viewProperties.ts
@@ -4,10 +4,10 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { IActionContext, openReadOnlyJson } from 'vscode-azureextensionui';
-import { SiteTreeItem } from '../explorer/SiteTreeItem';
-import { WebAppTreeItem } from '../explorer/WebAppTreeItem';
 import { ext } from '../extensionVariables';
 import { localize } from '../localize';
+import { SiteTreeItem } from '../tree/SiteTreeItem';
+import { WebAppTreeItem } from '../tree/WebAppTreeItem';
 
 export async function viewProperties(context: IActionContext, node?: SiteTreeItem): Promise<void> {
     if (!node) {
