+++ src/tree/NotAvailableTreeItem.ts
